Achieve maximum coverage(sort of-read below). This closes #23.
For some reason, scoverage shows some parts of the code as not being tested. I have confirmed they are indeed tested. I believe this may be due
to the liberal usage of anon functions within the tests; as I cannot properly debug them I am left in the dark here.