Measurements now sorted on insertion. This closes #31.
However it is quite inefficient. It is a simple approach where, whenever a measurement is added, the measurements are sorted.
This is O(n^2logn) so with sufficient measurements things will slow down a lot. Ideally I would implement this with SortedSet or PriorityQueues but I ran into
a lot of problems with the latter approach. As time is of the essence this, for now, will have to suffice.