Measurements are now sorted only when exporting them into a file.
This commit reverts the previous one that tackled the issue of sorting:
instead of sorting the measurements on insertion, these are only sorted when exported to a file.
This implies that the state of the program will not alter and I believe is a more elegant and efficient approach,
seeing as the cost of sorting is only paid for when exporting the measurements which doesn't happen very often.